LPS-121259 Better to not use a hook it is not reusable
