+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.js
@@ -15,7 +15,6 @@
 import PropTypes from 'prop-types';
 import React, {useEffect} from 'react';
 
-import useAutoExtendSession from '../../core/hooks/useAutoExtendSession';
 import {StyleBookContextProvider} from '../../plugins/page-design-options/hooks/useStyleBook';
 import {INIT} from '../actions/types';
 import {config} from '../config/index';
@@ -31,10 +30,23 @@ import Sidebar from './Sidebar';
 import Toolbar from './Toolbar';
 import URLParser from './URLParser';
 
+const DEFAULT_SESSION_LENGTH = 60 * 1000;
+
 export default function App({state}) {
 	const initialState = reducer(state, {type: INIT});
 
-	useAutoExtendSession();
+	useEffect(() => {
+		if (Liferay.Session && config.autoExtendSessionEnabled) {
+			const sessionLength =
+				Liferay.Session.get('sessionLength') || DEFAULT_SESSION_LENGTH;
+
+			const interval = setInterval(() => {
+				Liferay.Session.extend();
+			}, sessionLength / 2);
+
+			return () => clearInterval(interval);
+		}
+	}, []);
 
 	return (
 		<StoreContextProvider initialState={initialState} reducer={reducer}>
